ADD: literate the GC2.CS class

+++ LNativeMemory/IAllocator.cs
@@ -11,7 +11,7 @@ namespace LNativeMemory
 
         void Reset();
 
-        uint BytesLeft { get; }
-        uint TotalBytes { get; }
+        long BytesLeft { get; }
+        long TotalBytes { get; }
     }
 }
